[-] BO : Fix KPI refresh functions
In kpi.tpl the refresh on page load function was taken out because the
data is already loaded from the HelperKpi class with the variable
'source'. Moved the 'refresh' conditions up so the script isn't loaded
in unnecessarily.

In admin.js checked if the particular refresh function for the kpi box
exists before calling it.