FO: Fixed canonical redirects
Fixed canonical redirects for manufacturers and suppliers list pages.
Was only working if the object of the supplier or manufacturer was
loaded, which isn't the case on the lists.

parent::init was moved to the bottom of the init functions so that the
manufacturer's and supplier's cannonical redirect would fire before the
front controller's.