Js responsive value is not changed on resize (#7704)
* FO: Js responsive value is not changed on resize

To corectly calculate prestashop.responsive.mobile value it is needed to base on actual window width. Previosly it was base on _cw value which stored window with value calculated before window resize.

*  FO: Js responsive value is not changed on resize

To corectly calculate prestashop.responsive.mobile value it is needed to base on actual window width. Previosly it was base on _cw value which stored window with value calculated before window resize.