Make sure the alias hook function is called if the hook was registered to an alias
Before this change, when `someHook` was invoked, modules having registered to `someHookAlias`
wouldn't be notified, because Hook would call `hookSomeHook()` (the method matching the
canonical hook) instead of `hookSomeHookAlias()` (the method matching the aliased hook,
present in the module). This change ensures that modules hooking to an alias receive the
appropriate method call.